pyproject.toml
[project]
name = "agent-howto"
version = "0.1.0"
description = "A uv-based learning roadmap for building agents from demos."
authors = [
    { name = "yu.xia", email = "yuxianice@163.com" }
]
requires-python = ">=3.14"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "socksio>=1.0.0",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
    "types-requests>=2.33.0.20260408",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
mypy_path = "hello-agent/chapter1/first_agent"
